one_template/requirements.txt
attrdict==2.0.1
loguru==0.6.0
nni==2.10
numpy==1.23.4
pandas==1.5.1
PyYAML==6.0
scikit_learn==1.1.3
spacy==3.4.2
torch>=1.8.1
tqdm==4.64.1
python>=3.7